I got a wrong result, but it is result, fix bugs from now on.

+++ src/com/danliu/stock/model/NotTraded.java
@@ -39,7 +39,7 @@ public class NotTraded extends Trade {
     }
 
     @Override
-    public float getTradeCount() {
+    public int getTradeCount() {
         throw new IllegalAccessError("not supported.");
     }
